requirements.txt
Django==4.2.7
Pillow==10.1.0
stripe==7.6.0
python-decouple==3.8
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-allauth==0.57.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
